SQLMesh Production Fixes (#2936)
* fix: add `dt` boundaries for bigquery storage api calls

* use date

* ensure issue number is int64

* fix type casting

* Explicitly set kube context

* use larger batch size

* improve partition management

* Change scaling policy

* more partitioning fixes

* more partitioning fixes

* fix model

* fix: missing comma

* fix: fixes openrank query date usage

+++ warehouse/metrics_mesh/models/staging/github/stg_github__pull_requests.sql
@@ -19,7 +19,7 @@ select
   pre.actor.login as actor_login,
   CONCAT('PULL_REQUEST_', UPPER(json_extract_string(pre.payload, '$.action')))
     as "type",
-  json_extract(pre.payload, '$.number')::INT as "number",
+  json_extract(pre.payload, '$.number')::BIGINT as "number",
   strptime(
     json_extract_string(pre.payload, '$.pull_request.created_at'),
     '%Y-%m-%dT%H:%M:%SZ'
